+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
+## [0.2.3] - 2026-07-03
+
 ### Changed
 
 - Compile results are correlated EXACTLY via the session's new `project-ack`